build: resolve 1 instance of -Wnarrowing
lib/freebusy.cpp: In function 'bool find_recur_times':
lib/freebusy.cpp:247:47: warning: narrowing conversion of
'(((long unsigned int)ut) + ((long unsigned int)((((unsigned
int)apr.APPOINTMENT_RECUR_PAT::endtimeoffset) - ((unsigned
int)apr.APPOINTMENT_RECUR_PAT::starttimeoffset)) * 60)))' from 'long
unsigned int' to 'time_t' {aka 'long int'} [-Wnarrowing]

+++ lib/freebusy.cpp
@@ -244,7 +244,7 @@ static bool find_recur_times(const ical_component *tzcom,
 		};
 		if (std::any_of(&ei[0], &ei[apr.exceptioncount], time_test))
 			continue;
-		evlist.push_back(event{ut, ut + (apr.endtimeoffset - apr.starttimeoffset) * 60});
+		evlist.push_back(event{ut, ut + static_cast<long>((apr.endtimeoffset - apr.starttimeoffset) * 60)});
 		if (ut >= end_time)
 			break;
 	} while (irrule.iterate());
